Player object can now be removed from the game and dependendts will listen for it. Added Player::setState, but physics might not obey it...
GameTestThing now handles the bare minimum of the
"Setup world" portion of the server control flow diagram.
(i.e creates a local player and adds it to the world, and
sets it as the local player)
Also listens for playerCreated events from the network
and adds the player to the world when they happen.

NetworkTestStuff::clientConnected will create a player
and signal that it has done so, then register the object
(still doesn't send world state or tell the client which
player they are)